config: zynq-common: Remove compatible env setting
This patch removes the extra compatible setting and only leaves the old
method for changing the IIO devicei, attr_name and attr_val.

The user will use the combination of these two commands:

fw_setenv attr_name compatible
fw_setenv attr_val ad9364

This will ensure that we don't change anything to the documentation that is
already in place for Pluto.

Signed-off-by: Mircea Caprioru <mircea.caprioru@analog.com>